8329494: Serial: Merge GenMarkSweep into MarkSweep
