+++ make/hotspot/lib/JvmFeatures.gmk
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2013, 2023, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2013, 2024, Oracle and/or its affiliates. All rights reserved.
 # DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
 #
 # This code is free software; you can redistribute it and/or modify it
@@ -214,7 +214,6 @@ ifeq ($(call check-jvm-feature, opt-size), true)
       frame_x86.cpp \
       genCollectedHeap.cpp \
       generation.cpp \
-      genMarkSweep.cpp \
       growableArray.cpp \
       handles.cpp \
       hashtable.cpp \
