--- src/hotspot/share/gc/serial/genMarkSweep.hpp
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2001, 2019, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- *
- */
-
-#ifndef SHARE_GC_SERIAL_GENMARKSWEEP_HPP
-#define SHARE_GC_SERIAL_GENMARKSWEEP_HPP
-
-#include "gc/serial/markSweep.hpp"
-
-class GenMarkSweep : public MarkSweep {
- public:
-  static void invoke_at_safepoint(bool clear_all_softrefs);
-
- private:
-  // Mark live objects
-  static void phase1_mark(bool clear_all_softrefs);
-
-  // Temporary data structures for traversal and storing/restoring marks
-  static void allocate_stacks();
-  static void deallocate_stacks();
-};
-
-#endif // SHARE_GC_SERIAL_GENMARKSWEEP_HPP
